feat(#29): implement second-survey-screen and second-questio-widget

+++ lib/models/second_survey/bloc/second_survey_event.dart
@@ -5,9 +5,10 @@ abstract class SecondSurveyEvent {}
 class EventChangeValue extends SecondSurveyEvent {
   String key;
   int index;
-  int value;
+  double value;
 
-  EventChangeValue({required this.key,required this.index,required this.value});
+  EventChangeValue(
+      {required this.key, required this.index, required this.value});
 }
 
 class EventLoadSecondSurvey extends SecondSurveyEvent {
